feat: Add quest Lesson Series (#86)
* feat: Lesson Quest Series

Add 4 new quests, one for each character in the base game, that appears when you are not playing the character. When taken, the player obtains a Questbound Clumsy-Offering/Nightmare/Reboot/Blasphemy with ethereal and all upsides removed. After being played 4/4/10/3 times, the Clumsy card is removed and replaced with an upgraded version of the original card it is based on.

* fix: Lesson Quest Series comments

whoops